Move duplicated fixtures to common conftest.py
